+++ packages/typespec-rust/CHANGELOG.md
@@ -86,6 +86,7 @@
 
 * Added support for `plainDate` and `plainTime` types. They're emitted as `String` types.
 * Added support for the `safeint` type. It's emitted as a `serde_json::Number` type.
+* Added support for optional path parameters.
 
 ### Bugs Fixed
 
@@ -93,6 +94,10 @@
 * Fixed bad codegen for model with literal values.
 * Fixed incorrect behavior for numeric types that use string encoding.
 * Fixed decimal types to properly handle string/float encodings.
+
+### Other Changes
+
+* Required path parameters that are empty will return an error.
 
 ## 0.14.1 (2025-05-07)
 
@@ -123,7 +128,6 @@
 
 * Added support for pageable methods that use a continuation token when fetching pages.
 * Added support for TypeSpec `decimal` and `decimal128` types.
-* Added support for optional path parameters.
 
 ### Bugs Fixed
 
@@ -134,7 +138,6 @@
 
 * Errors in the emitter are no longer surfaced as a crash.
 * Skip `cargo fmt` if the emitter fails.
-* Required path parameters that are empty will return an error.
 
 ## 0.13.3 (2025-04-04)